+++ PACECloudSDK/AppKit/AppManagement/OneTimeLocationProvider.swift
@@ -16,7 +16,8 @@ class OneTimeLocationProvider: NSObject, CLLocationManagerDelegate {
     weak var delegate: OneTimeLocationProviderDelegate?
 
     private let manager: CLLocationManager = .init()
-    private var locationUpdateHandler: ((CLLocation?) -> Void)?
+    private var locationUpdateHandler: [((CLLocation?) -> Void)] = .init()
+    private lazy var workQueue: DispatchQueue = .init(label: "location-provider-queue")
 
     override init() {
         super.init()
@@ -24,20 +25,34 @@ class OneTimeLocationProvider: NSObject, CLLocationManagerDelegate {
         manager.delegate = self
     }
 
-    func requestLocation(locationUpdateHandler: ((CLLocation?) -> Void)? = nil) {
-        self.locationUpdateHandler = locationUpdateHandler
+    func requestLocation(completion: ((CLLocation?) -> Void)? = nil) {
+        workQueue.async { [weak self] in
+            if let handler = completion {
+                self?.locationUpdateHandler.append(handler)
+            }
+        }
+
         manager.requestLocation()
     }
 
     func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
         let location = locations.last
         delegate?.didFinishLocationRequest(with: location)
-        locationUpdateHandler?(location)
+
+        notifyHandlers(location)
     }
 
     func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
         AppKitLogger.e("[OneTimeLocationProvider] Requesting location failed with error \(error)")
         delegate?.didFinishLocationRequest(with: nil)
-        locationUpdateHandler?(nil)
+
+        notifyHandlers(nil)
+    }
+
+    private func notifyHandlers(_ location: CLLocation?) {
+        workQueue.async { [weak self] in
+            self?.locationUpdateHandler.forEach { $0(location) }
+            self?.locationUpdateHandler.removeAll()
+        }
     }
 }